refactor(app): move apps:info to CLI & upgrade to oclif (#2524)
* Consolidate unwrap

* Add info.ts and info.unit.test.ts files

* Update description, topic, alias, examples, flags & args

* Move old logic over for conversion

* Update imports

* WIP converting types

* Finish type conversion

* Migrate formatDate function

* WIP apps:info initially working

* WIP working version of command

Use @ts-ignore to stop typescript yelling. Will have to resolve this later once all variations of command successfully work.

* Fix appExtendedResponse. Add check to extended flag.

* Successfully have command working with all flags

* WIP get first test working

* refactor: ~half the tests to core & TS

* refactor: upgrade tests to TS & core

* clean up

* fix: type correction. Delete v5 apps:info

* tweak

---------
