+++ packages/cli/package.json
@@ -49,6 +49,7 @@
     "debug": "4.1.1",
     "edit-string": "^1.1.6",
     "execa": "5.1.1",
+    "filesize": "^4.0.0",
     "foreman": "^3.0.1",
     "fs-extra": "7.0.1",
     "github-url-to-object": "^4.0.4",
